Add a simple callback func
