+++ pkg/controller/constants.go
@@ -1,0 +1,15 @@
+package controller
+
+const (
+	FrameworkName = "FrameworkName"
+	Name          = "Name"
+	Namespace     = "Namespace"
+	Params        = "Params"
+)
+
+// Error types
+
+const (
+	Warning string = "Warning"
+)
+
