+++ pkg/controller/planexecution/planexecution_controller.go
@@ -19,21 +19,40 @@ import (
 	"context"
 	"fmt"
 	"log"
+	"strconv"
 
-	kudov1alpha1 "github.com/kudobuilder/kudo/pkg/apis/kudo/v1alpha1"
-	"github.com/kudobuilder/kudo/pkg/engine"
 	"github.com/kudobuilder/kudo/pkg/util/health"
+	"github.com/kudobuilder/kudo/pkg/util/template"
+	"gopkg.in/yaml.v2"
 	appsv1 "k8s.io/api/apps/v1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/api/extensions/v1beta1"
 	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+	"sigs.k8s.io/controller-runtime/pkg/event"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+	"sigs.k8s.io/controller-runtime/pkg/predicate"
+	"sigs.k8s.io/kustomize/k8sdeps/kunstruct"
+	"sigs.k8s.io/kustomize/k8sdeps/transformer"
+	"sigs.k8s.io/kustomize/pkg/fs"
+	"sigs.k8s.io/kustomize/pkg/loader"
+	"sigs.k8s.io/kustomize/pkg/patch"
+	"sigs.k8s.io/kustomize/pkg/resmap"
+	"sigs.k8s.io/kustomize/pkg/resource"
+	"sigs.k8s.io/kustomize/pkg/target"
+	ktypes "sigs.k8s.io/kustomize/pkg/types"
+
+	kudov1alpha1 "github.com/kudobuilder/kudo/pkg/apis/kudo/v1alpha1"
+
+	"github.com/kudobuilder/kudo/pkg/engine"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 	"sigs.k8s.io/controller-runtime/pkg/source"
 )
@@ -75,9 +94,9 @@ func add(mgr manager.Manager, r reconcile.Reconciler) error {
 		err = c.Watch(
 			&source.Kind{Type: e},
 			&handler.EnqueueRequestsFromMapFunc{
-				ToRequests: engine.ReconcileRequestsMapperFunc(mgr),
+				ToRequests: reconcileRequestsMapperFunc(mgr),
 			},
-			engine.PlanEventPredicateFunc())
+			planEventPredicateFunc())
 		if err != nil {
 			return err
 		}
@@ -104,7 +123,7 @@ type ReconcilePlanExecution struct {
 // +kubebuilder:rbac:groups="",resources=events;configmaps,verbs=get;list;watch;create;patch
 // +kubebuilder:rbac:groups=policy,resources=poddisruptionbudgets;poddisruptionbudgets.policy,verbs=get;list;watch;create;update;patch;delete
 func (r *ReconcilePlanExecution) Reconcile(request reconcile.Request) (reconcile.Result, error) {
-	// Fetch the PlanExecution instance
+	// Fetch the PlanExecution Instance
 	planExecution := &kudov1alpha1.PlanExecution{}
 	err := r.Get(context.TODO(), request.NamespacedName, planExecution)
 	if err != nil {
@@ -118,7 +137,7 @@ func (r *ReconcilePlanExecution) Reconcile(request reconcile.Request) (reconcile
 		return reconcile.Result{Requeue: true}, err
 	}
 
-	//Get Instance Object
+	// Get Instance Object
 	instance := &kudov1alpha1.Instance{}
 	err = r.Get(context.TODO(),
 		types.NamespacedName{
@@ -127,9 +146,9 @@ func (r *ReconcilePlanExecution) Reconcile(request reconcile.Request) (reconcile
 		},
 		instance)
 	if err != nil {
-		// Can't find the instance. Update status.
+		// Can't find the Instance. Update status.
 		planExecution.Status.State = kudov1alpha1.PhaseStateError
-		r.recorder.Event(planExecution, "Warning", "InvalidInstance", fmt.Sprintf("Could not find required instance (%v)", planExecution.Spec.Instance.Name))
+		r.recorder.Event(planExecution, "Warning", "InvalidInstance", fmt.Sprintf("Could not find required Instance (%v)", planExecution.Spec.Instance.Name))
 		log.Printf("PlanExecutionController: Error getting Instance %v in %v: %v",
 			planExecution.Spec.Instance.Name,
 			planExecution.Spec.Instance.Namespace,
@@ -137,7 +156,7 @@ func (r *ReconcilePlanExecution) Reconcile(request reconcile.Request) (reconcile
 		return reconcile.Result{}, err
 	}
 
-	//See if this has already been processed
+	// See if this has already been processed
 	if planExecution.Status.State == kudov1alpha1.PhaseStateComplete {
 		log.Printf("PlanExecutionController: PlanExecution \"%v\" has already run to completion, not processing.", planExecution.Name)
 		return reconcile.Result{}, nil
@@ -164,7 +183,7 @@ func (r *ReconcilePlanExecution) Reconcile(request reconcile.Request) (reconcile
 	err = r.Update(context.TODO(), instance)
 	if err != nil {
 		r.recorder.Event(planExecution, "Warning", "UpdateError", fmt.Sprintf("Could not update the ActivePlan for (%v): %v", planExecution.Spec.Instance.Name, err))
-		log.Printf("PlanExecutionController: Update of instance with ActivePlan errored: %v", err)
+		log.Printf("PlanExecutionController: Update of Instance with ActivePlan errored: %v", err)
 	}
 
 	//Get associated FrameworkVersion
@@ -186,17 +205,20 @@ func (r *ReconcilePlanExecution) Reconcile(request reconcile.Request) (reconcile
 		return reconcile.Result{}, err
 	}
 
-	configs, err := engine.ParseConfig(planExecution, instance, frameworkVersion, r.recorder)
+	configs, err := engine.ParseConfig(instance, frameworkVersion, func(eventtype, reason, message string) {
+		r.recorder.Event(planExecution, eventtype, reason, message)
+	})
 	if err != nil {
+		log.Printf("PlanExecutionController: %v", err)
 		return reconcile.Result{}, err
 	}
 
-	//Get Plan from FrameworkVersion:
-	//Right now must match exactly.  In the future have defaults/backups:
-	// e.g. if no "upgrade", call "update"
-	// if no "update" call "deploy"
-	// When we have this we'll have to keep the active plan in the status since
-	// that might not match the "requested" plan.
+	// Get Plan from FrameworkVersion:
+	// Right now must match exactly.  In the future have defaults/backups:
+	//  e.g. if no "upgrade", call "update"
+	//  if no "update" call "deploy"
+	//  When we have this we'll have to keep the active plan in the status since
+	//  that might not match the "requested" plan.
 	executedPlan, ok := frameworkVersion.Spec.Plans[planExecution.Spec.PlanName]
 	if !ok {
 		r.recorder.Event(planExecution, "Warning", "InvalidPlan", fmt.Sprintf("Could not find required plan (%v)", planExecution.Spec.PlanName))
@@ -208,12 +230,12 @@ func (r *ReconcilePlanExecution) Reconcile(request reconcile.Request) (reconcile
 	planExecution.Status.Name = planExecution.Spec.PlanName
 	planExecution.Status.Strategy = executedPlan.Strategy
 
-	err = engine.PopulatePlanExecutionPhases(basePath, &executedPlan, planExecution, instance, frameworkVersion, configs, r.recorder)
+	err = PopulatePlanExecutionPhases(basePath, &executedPlan, planExecution, instance, frameworkVersion, configs, r.recorder)
 	if err != nil {
 		return reconcile.Result{}, err
 	}
 
-	err = engine.RunPhases(&executedPlan, planExecution, instance, r.Client, r.scheme)
+	err = RunPhases(&executedPlan, planExecution, instance, r.Client, r.scheme)
 	if err != nil {
 		return reconcile.Result{}, err
 	}
@@ -229,7 +251,7 @@ func (r *ReconcilePlanExecution) Reconcile(request reconcile.Request) (reconcile
 	instance.Status.Status = planExecution.Status.State
 	err = r.Client.Update(context.TODO(), instance)
 	if err != nil {
-		log.Printf("Error updating instance status to %v: %v\n", instance.Status.Status, err)
+		log.Printf("Error updating Instance status to %v: %v\n", instance.Status.Status, err)
 	}
 
 	return reconcile.Result{}, nil
@@ -278,3 +300,383 @@ func (r *ReconcilePlanExecution) Cleanup(obj runtime.Object) error {
 
 	return nil
 }
+
+// Watch for Deployments, Jobs and StatefulSets
+// Define a mapping from the object in the event to one or more
+// objects to Reconcile.  Specifically this calls for
+// a reconciliation of any objects "Owner".
+func reconcileRequestsMapperFunc(mgr manager.Manager) handler.ToRequestsFunc {
+	return func(a handler.MapObject) []reconcile.Request {
+		owners := a.Meta.GetOwnerReferences()
+		requests := make([]reconcile.Request, 0)
+		for _, owner := range owners {
+			// if owner is an Instance, we also want to queue up the
+			// PlanExecution in the Status section
+			inst := &kudov1alpha1.Instance{}
+			err := mgr.GetClient().Get(context.TODO(), client.ObjectKey{
+				Name:      owner.Name,
+				Namespace: a.Meta.GetNamespace(),
+			}, inst)
+
+			if err != nil {
+				log.Printf("Error getting Instance object: %v", err)
+			} else {
+				log.Printf("Adding \"%v\" to reconcile", inst.Status.ActivePlan.Name)
+				requests = append(requests, reconcile.Request{
+					NamespacedName: types.NamespacedName{
+						Name:      inst.Status.ActivePlan.Name,
+						Namespace: inst.Status.ActivePlan.Namespace,
+					},
+				})
+			}
+		}
+		return requests
+	}
+}
+
+func planEventPredicateFunc() predicate.Funcs {
+	// 'UpdateFunc' and 'CreateFunc' used to judge if a event about the object is
+	// what we want. If that is true, the event will be processed by the reconciler.
+	// PlanExecutions should be mostly immutable.  Updates should only
+	msg := "PlanEventPredicate: Received update event for an Instance named: %v"
+	return predicate.Funcs{
+		UpdateFunc: func(e event.UpdateEvent) bool {
+			log.Printf(msg, e.MetaNew.GetName())
+			return e.ObjectOld != e.ObjectNew
+		},
+		CreateFunc: func(e event.CreateEvent) bool {
+			log.Printf(msg, e.Meta.GetName())
+			return true
+		},
+		DeleteFunc: func(e event.DeleteEvent) bool {
+			// TODO send event for Instance that plan was deleted
+			log.Printf(msg, e.Meta.GetName())
+			return true
+		},
+	}
+}
+
+func PopulatePlanExecutionPhases(basePath string, executedPlan *kudov1alpha1.Plan, planExecution *kudov1alpha1.PlanExecution, instance *kudov1alpha1.Instance, frameworkVersion *kudov1alpha1.FrameworkVersion, configs map[string]interface{}, recorder record.EventRecorder) error {
+	planExecution.Status.Phases = make([]kudov1alpha1.PhaseStatus, len(executedPlan.Phases))
+	var err error
+	for i, phase := range executedPlan.Phases {
+		planExecution.Status.Phases[i].Name = phase.Name
+		planExecution.Status.Phases[i].Strategy = phase.Strategy
+		planExecution.Status.Phases[i].State = kudov1alpha1.PhaseStatePending
+		planExecution.Status.Phases[i].Steps = make([]kudov1alpha1.StepStatus, len(phase.Steps))
+		for j, step := range phase.Steps {
+			// fetch FrameworkVersion
+			// get the task name from the step
+			// get the task definition from the FV
+			// create the kustomize templates
+			// apply
+			configs["PlanName"] = planExecution.Spec.PlanName
+			configs["PhaseName"] = phase.Name
+			configs["StepName"] = step.Name
+			configs["StepNumber"] = strconv.FormatInt(int64(j), 10)
+
+			var objs []runtime.Object
+			engine := engine.New()
+
+			for _, t := range step.Tasks {
+				// resolve task
+				if taskSpec, ok := frameworkVersion.Spec.Tasks[t]; ok {
+					var resources []string
+					fsys := fs.MakeFakeFS()
+
+					for _, res := range taskSpec.Resources {
+						if resource, ok := frameworkVersion.Spec.Templates[res]; ok {
+							templatedYaml, err := engine.Render(resource, configs)
+							if err != nil {
+								recorder.Event(planExecution, "Warning", "InvalidPlanExecution", fmt.Sprintf("Error expanding template: %v", err))
+								log.Printf("PlanExecutionController: Error expanding template: %v", err)
+							}
+							fsys.WriteFile(fmt.Sprintf("%s/%s", basePath, res), []byte(templatedYaml))
+							resources = append(resources, res)
+
+						} else {
+							recorder.Event(planExecution, "Warning", "InvalidPlanExecution", fmt.Sprintf("Error finding resource named %v for framework version %v", res, frameworkVersion.Name))
+							log.Printf("PlanExecutionController: Error finding resource named %v for framework version %v", res, frameworkVersion.Name)
+							return err
+						}
+					}
+
+					kustomization := &ktypes.Kustomization{
+						NamePrefix: instance.Name + "-",
+						Namespace:  instance.Namespace,
+						CommonLabels: map[string]string{
+							"heritage":      "kudo",
+							"app":           frameworkVersion.Spec.Framework.Name,
+							"version":       frameworkVersion.Spec.Version,
+							"Instance":      instance.Name,
+							"planexecution": planExecution.Name,
+							"plan":          planExecution.Spec.PlanName,
+							"phase":         phase.Name,
+							"step":          step.Name,
+						},
+						GeneratorOptions: &ktypes.GeneratorOptions{
+							DisableNameSuffixHash: true,
+						},
+						Resources:             resources,
+						PatchesStrategicMerge: []patch.StrategicMerge{},
+					}
+
+					yamlBytes, err := yaml.Marshal(kustomization)
+					if err != nil {
+						return err
+					}
+
+					fsys.WriteFile(fmt.Sprintf("%s/kustomization.yaml", basePath), yamlBytes)
+
+					ldr, err := loader.NewLoader(basePath, fsys)
+					if err != nil {
+						return err
+					}
+					defer ldr.Cleanup()
+
+					rf := resmap.NewFactory(resource.NewFactory(kunstruct.NewKunstructuredFactoryImpl()))
+					kt, err := target.NewKustTarget(ldr, fsys, rf, transformer.NewFactoryImpl())
+					if err != nil {
+						return err
+					}
+
+					allResources, err := kt.MakeCustomizedResMap()
+					if err != nil {
+						return err
+					}
+
+					res, err := allResources.EncodeAsYaml()
+					if err != nil {
+						return err
+					}
+
+					objsToAdd, err := template.ParseKubernetesObjects(string(res))
+					if err != nil {
+						recorder.Event(planExecution, "Warning", "InvalidPlanExecution", fmt.Sprintf("Error creating Kubernetes objects from step %v in phase %v of plan %v: %v", step.Name, phase.Name, planExecution.Name, err))
+						log.Printf("PlanExecutionController: Error creating Kubernetes objects from step %v in phase %v of plan %v: %v", step.Name, phase.Name, planExecution.Name, err)
+						return err
+					}
+					objs = append(objs, objsToAdd...)
+				} else {
+					recorder.Event(planExecution, "Warning", "InvalidPlanExecution", fmt.Sprintf("Error finding task named %s for framework version %s", taskSpec, frameworkVersion.Name))
+					log.Printf("PlanExecutionController: Error finding task named %s for framework version %s", taskSpec, frameworkVersion.Name)
+					return nil
+				}
+			}
+
+			planExecution.Status.Phases[i].Steps[j].Name = step.Name
+			planExecution.Status.Phases[i].Steps[j].Objects = objs
+			planExecution.Status.Phases[i].Steps[j].Delete = step.Delete
+			log.Printf("PlanExecutionController: Phase \"%v\" Step \"%v\" has %v object(s)", phase.Name, step.Name, len(objs))
+		}
+	}
+	return nil
+}
+
+func MutateFn(oldObj runtime.Object) controllerutil.MutateFn {
+	return func(newObj runtime.Object) error {
+		//TODO Clean this up.  I don't like having to do a switch here
+		switch t := newObj.(type) {
+		case *appsv1.StatefulSet:
+			log.Printf("PlanExecutionController: CreateOrUpdate: StatefulSet %+v", t.Name)
+
+			newSs := newObj.(*appsv1.StatefulSet)
+			ss, ok := oldObj.(*appsv1.StatefulSet)
+			if !ok {
+				return fmt.Errorf("object passed in doesn't match expected StatefulSet type")
+			}
+
+			// We need some specialized logic in there.  We can't just copy the Spec since there are other values
+			// like spec.updateState, spec.volumeClaimTemplates, etc that are all
+			// generated from the object by the k8s controller.  We just want to update things we can change
+			newSs.Spec.Replicas = ss.Spec.Replicas
+
+			return nil
+		case *appsv1.Deployment:
+			newD := newObj.(*appsv1.Deployment)
+			d, ok := oldObj.(*appsv1.Deployment)
+			if !ok {
+				return fmt.Errorf("object passed in doesn't match expected deployment type")
+			}
+			newD.Spec.Replicas = d.Spec.Replicas
+			return nil
+		case *v1beta1.Deployment:
+			newD := newObj.(*v1beta1.Deployment)
+			d, ok := oldObj.(*v1beta1.Deployment)
+			if !ok {
+				return fmt.Errorf("object passed in doesn't match expected deployment type")
+			}
+			newD.Spec.Replicas = d.Spec.Replicas
+			return nil
+
+		case *batchv1.Job:
+			// job := oldObj.(*batchv1.Job)
+
+		case *kudov1alpha1.Instance:
+			// i := oldObj.(*kudov1alpha1.Instance)
+
+		//unless we build logic for what a healthy object is, assume its healthy when created
+		default:
+			log.Print("PlanExecutionController: CreateOrUpdate: Type is not implemented yet")
+			return nil
+		}
+
+		return nil
+
+	}
+}
+
+func Cleanup(c client.Client, obj runtime.Object) error {
+	switch obj := obj.(type) {
+	case *batchv1.Job:
+		//We need to see if there's a current job on the system that matches this exactly (with labels)
+		log.Printf("PlanExecutionController.Cleanup: *batchv1.Job %v", obj.Name)
+
+		present := &batchv1.Job{}
+		key, _ := client.ObjectKeyFromObject(obj)
+		err := c.Get(context.TODO(), key, present)
+		if errors.IsNotFound(err) {
+			//this is fine, its good to go
+			log.Printf("PlanExecutionController: Could not find job \"%v\" in cluster. Good to make a new one.", key)
+			return nil
+		}
+		if err != nil {
+			//Something else happened
+			return err
+		}
+		//see if the job in the cluster has the same labels as the one we're looking to add.
+		for k, v := range obj.Labels {
+			if v != present.Labels[k] {
+				//need to delete the present job since its got labels that aren't the same
+				log.Printf("PlanExecutionController: Different values for job key \"%v\": \"%v\" and \"%v\"", k, v, present.Labels[k])
+				err = c.Delete(context.TODO(), present)
+				return err
+			}
+		}
+		for k, v := range present.Labels {
+			if v != obj.Labels[k] {
+				//need to delete the present job since its got labels that aren't the same
+				log.Printf("PlanExecutionController: Different values for job key \"%v\": \"%v\" and \"%v\"", k, v, obj.Labels[k])
+				err = c.Delete(context.TODO(), present)
+				return err
+			}
+		}
+		return nil
+	}
+
+	return nil
+}
+
+func RunPhases(executedPlan *kudov1alpha1.Plan, planExecution *kudov1alpha1.PlanExecution, instance *kudov1alpha1.Instance, c client.Client, scheme *runtime.Scheme) error {
+	var err error
+	for i, phase := range planExecution.Status.Phases {
+		//If we still want to execute phases in this plan
+		//check if phase is healthy
+		for j, s := range phase.Steps {
+			planExecution.Status.Phases[i].Steps[j].State = kudov1alpha1.PhaseStateComplete
+
+			for _, obj := range s.Objects {
+				if s.Delete {
+					log.Printf("PlanExecutionController: Step \"%v\" was marked to delete object %+v", s.Name, obj)
+					err = c.Delete(context.TODO(), obj, client.PropagationPolicy(metav1.DeletePropagationForeground))
+					if errors.IsNotFound(err) || err == nil {
+						//This is okay
+						log.Printf("PlanExecutionController: Object was already deleted or did not exist in step \"%v\"", s.Name)
+					}
+					if err != nil {
+						log.Printf("PlanExecutionController: Error deleting object in step \"%v\": %v", s.Name, err)
+						planExecution.Status.Phases[i].State = kudov1alpha1.PhaseStateError
+						planExecution.Status.Phases[i].Steps[j].State = kudov1alpha1.PhaseStateError
+						return err
+					}
+					continue
+				}
+
+				// Make sure this object is applied to the cluster. Get back the Instance from
+				// the cluster so we can see if it's healthy or not
+				if err = controllerutil.SetControllerReference(instance, obj.(metav1.Object), scheme); err != nil {
+					return err
+				}
+
+				//Some objects don't update well.  We capture the logic here to see if we need to cleanup the current object
+				err = Cleanup(c, obj)
+				if err != nil {
+					log.Printf("PlanExecutionController: Cleanup failed: %v", err)
+				}
+
+				arg := obj.DeepCopyObject()
+				result, err := controllerutil.CreateOrUpdate(context.TODO(), c, arg, MutateFn(obj))
+
+				if err != nil {
+					log.Printf("PlanExecutionController: Error CreateOrUpdate Object in step \"%v\": %v", s.Name, err)
+					planExecution.Status.Phases[i].State = kudov1alpha1.PhaseStateError
+					planExecution.Status.Phases[i].Steps[j].State = kudov1alpha1.PhaseStateError
+
+					return err
+				}
+				log.Printf("PlanExecutionController: CreateOrUpdate resulted in: %v", result)
+
+				// get the existing object meta
+				metaObj := obj.(metav1.Object)
+
+				// retrieve the existing object
+				key := client.ObjectKey{
+					Name:      metaObj.GetName(),
+					Namespace: metaObj.GetNamespace(),
+				}
+
+				err = c.Get(context.TODO(), key, obj)
+
+				if err != nil {
+					log.Printf("PlanExecutionController: Error getting new object in step \"%v\": %v", s.Name, err)
+					planExecution.Status.Phases[i].State = kudov1alpha1.PhaseStateError
+					planExecution.Status.Phases[i].Steps[j].State = kudov1alpha1.PhaseStateError
+					return err
+				}
+				err = health.IsHealthy(c, obj)
+				if err != nil {
+					log.Printf("PlanExecutionController: Obj is NOT healthy: %+v", obj)
+					planExecution.Status.Phases[i].Steps[j].State = kudov1alpha1.PhaseStateInProgress
+					planExecution.Status.Phases[i].State = kudov1alpha1.PhaseStateInProgress
+				}
+			}
+			log.Printf("PlanExecutionController: Phase \"%v\" has strategy %v", phase.Name, phase.Strategy)
+			if phase.Strategy == kudov1alpha1.Serial {
+				//we need to skip the rest of the steps if this step is unhealthy
+				log.Printf("PlanExecutionController: Phase \"%v\" marked as serial", phase.Name)
+				if planExecution.Status.Phases[i].Steps[j].State != kudov1alpha1.PhaseStateComplete {
+					log.Printf("PlanExecutionController: Step \"%v\" isn't complete, skipping rest of steps in phase until it is", planExecution.Status.Phases[i].Steps[j].Name)
+					break //break step loop
+				} else {
+					log.Printf("PlanExecutionController: Step \"%v\" is healthy, so I can continue on", planExecution.Status.Phases[i].Steps[j].Name)
+				}
+			}
+
+			log.Printf("PlanExecutionController: Looked at step \"%v\"", s.Name)
+		}
+		if health.IsPhaseHealthy(planExecution.Status.Phases[i]) {
+			log.Printf("PlanExecutionController: Phase \"%v\" marked as healthy", phase.Name)
+			planExecution.Status.Phases[i].State = kudov1alpha1.PhaseStateComplete
+			continue
+		}
+
+		//This phase isn't quite ready yet.  Lets see what needs to be done
+		planExecution.Status.Phases[i].State = kudov1alpha1.PhaseStateInProgress
+
+		//Don't keep going to other plans if we're flagged to perform the phases in serial
+		if executedPlan.Strategy == kudov1alpha1.Serial {
+			log.Printf("PlanExecutionController: Phase \"%v\" not healthy, and plan marked as serial, so breaking.", phase.Name)
+			break
+		}
+		log.Printf("PlanExecutionController: Looked at phase \"%v\"", phase.Name)
+	}
+	return nil
+}
+
+type FIPP struct {
+	FrameworkVersion *kudov1alpha1.FrameworkVersion
+	Instance         *kudov1alpha1.Instance
+	PlanExecution    *kudov1alpha1.PlanExecution
+	Plan             *kudov1alpha1.Plan
+}
+
